[feature2] Pawn 클래스에 매개변수 없는 생성자 추가, 기본 생성자로 폰 생성시 흰색 폰 생성 테스트
