--- src/main/java/Pawn.java
@@ -1,11 +1,0 @@
-public class Pawn {
-    private final String color;
-
-    public Pawn(String color) {
-        this.color = color;
-    }
-
-    public String getColor() {
-        return color;
-    }
-}
