+++ src/main/java/pieces/Pawn.java
@@ -1,0 +1,20 @@
+package pieces;
+
+public class Pawn {
+    public static final String WHITE_COLOR = "white";
+    public static final String BLACK_COLOR = "black";
+
+    private final String color;
+
+    public Pawn() {
+        this.color = WHITE_COLOR;
+    }
+
+    public Pawn(String color) {
+        this.color = color;
+    }
+
+    public String getColor() {
+        return color;
+    }
+}
