+++ src/test/java/pieces/PawnTest.java
@@ -1,6 +1,9 @@
+package pieces;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import static org.junit.jupiter.api.Assertions.*;
 import static org.assertj.core.api.Assertions.*;
 
 public class PawnTest {
@@ -18,5 +21,12 @@ public class PawnTest {
     private void verifyPawn(final String color) {
         Pawn pawn = new Pawn(color);
         assertThat(pawn.getColor()).isEqualTo(color);
+    }
+
+    @Test
+    @DisplayName("기본 생성자로 폰 생성 시 흰색 폰이 생성되어야 한다")
+    public void create_기본생성자() throws Exception {
+        Pawn pawn = new Pawn();
+        assertEquals("white", pawn.getColor());
     }
 }
